Centralize doc comment handling for condense transform
